Switch markout_sobel_patches to an overapproximation
We know that for a given seam point at most two points can be
effected either side of it.

...#x##..
...##x##.
....##x##
.....##x#
The most extreme case

We can always assume that we affect all of these.
If we do assume this, then we can choose the range of points to
recalculate quite simply.

..##x##..
____^
..####..
____^
Cursor prior to and after removing the seam point

The bounds on this range is 'seam[y] - 2' to 'seam[y] + 1'.

After extensive profiling (I ran it 5 times before and after), it
does not seem to impact performance really at all (if anything my
intuition would be that it would improve by a slight margin), it
removes the need for the horrible type punning, and removes a
function definition.